% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{umu_palette}
\alias{umu_palette}
\title{A UmU palette generator}
\usage{
umu_palette(
  name = "original",
  n = 4,
  type = c("discrete", "continuous"),
  background = FALSE
)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{original},
\code{light},  \code{dark}. Default value is \code{original}.}

\item{n}{Number of colors desired. All palettes contain four colors. If
omitted, uses all colours. If \code{type} is continuous \code{n} specifies
the number of discrete colors to output. See \code{type}. Default value is
4 (all colors except background colors, see \code{background}.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}

\item{background}{Specifies if black and cream are included in the output.
Default value is \code{FALSE}.}
}
\value{
A vector of colours.
}
\description{
Three color palettes from the Umea University color profile. Link: https://www.aurora.umu.se/en/service-and-support/advice-and-guidelines/communication/visual-identity/
Based on the enjoyable wesanderson package.
}
\examples{
umu_palette("original")
umu_palette("light")
umu_palette("dark")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- umu_palette(21, name = "original", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
