% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umupalette.R
\docType{package}
\name{umupalette}
\alias{umupalette}
\title{umupalette}
\description{
list of palettes from Umea University's color identity.
}
\details{
list of palettes from Umea University's color identity.
}
